requirements.txt
# Data ingestion
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12

# Embeddings & clustering
sentence-transformers>=2.2
scikit-learn>=1.3
numpy>=1.24

# Database
psycopg2-binary>=2.9

# LLM APIs
groq>=0.4
google-generativeai>=0.5

# Schema validation
jsonschema>=4.20

# Config
python-dotenv>=1.0

# Testing
pytest>=7.4
